feat: init dynamic modules work
